Rename types/api.ts to apiClient.ts
