+++ src/screens/ConfirmTransferScreen.tsx
@@ -5,7 +5,8 @@ import { RootStackParamList, Transaction, TransactionStatus } from '../types';
 import { Button, Card, Loading } from '../components';
 import { COLORS, SPACING, FONT_SIZES } from '../constants';
 import { useAppDispatch, useAppSelector } from '../hooks';
-import { formatCurrency, generateTransactionId, Logger } from '../utils';
+import { formatCurrency, generateTransactionId } from '../utils';
+import { Logger } from '../utils/Logger';
 import { addTransaction } from '../store';
 import { updateBalance } from '../store';
 import { transferService } from "../services";
